Added timestamps to solr document
