+++ adsmp/solr_updater.py
@@ -1,6 +1,6 @@
 import requests
 import json
-from adsputils import setup_logging
+from adsputils import setup_logging, date2solrstamp
 import time
 
 
@@ -29,13 +29,35 @@ def get_orcid_claims(data, solrdoc):
     if 'unverified' in data:
         out['orcid_other'] = data['unverified']
     return out
+
+fmap = dict(metadata_mtime='bib_data_updated',
+           nonbib_mtime='nonbib_data_updated',
+           fulltext_mtime='fulltext_updated',
+           orcid_mtime='orcid_claims_updated',
+           metrics_mtime='metrics_updated'
+           )
+def get_timestamps(db_record, out):
+    out = {}
+    last_update = None
+    for k,v in fmap.items():
+        if v in db_record and db_record[v]:
+            t = db_record[v]
+            out[k] = date2solrstamp(t)
+            if last_update is None or t > last_update:
+                last_update = t
+    if last_update:
+        out['update_timestamp'] = date2solrstamp(last_update)
+    return out
      
 DB_COLUMN_DESTINATIONS = {
     'bib_data': '', 
     'orcid_claims': get_orcid_claims, 
     'nonbib_data': get_nonbib_for_solr,
     'id': 'id', 
-    'fulltext': 'body'}
+    'fulltext': 'body',
+    '#timestamps': get_timestamps, # use 'id' to be always called
+    
+    }
 
 
 def delete_by_bibcodes(bibcodes, urls):
@@ -103,13 +125,20 @@ def transform_json_record(db_record):
         if db_record.get(field, None):
             if target:
                 if callable(target):
-                    out.update(target(db_record.get(field), out)) # in the interest of speed, don't create copy of out
+                    x = target(db_record.get(field), out) # in the interest of speed, don't create copy of out
+                    if x:
+                        out.update(x) 
                 else:
                     out[target] = db_record.get(field)
             else:
                 if target is None:
                     continue
                 out.update(db_record.get(field))
+        elif field.startswith('#'):
+            if callable(target):
+                x = target(db_record, out) # in the interest of speed, don't create copy of out
+                if x:
+                    out.update(x) 
     
     return out
